==> 04. land cover.R <==
# Load package
library(terra)

# Clear environment 
rm(list = ls())

# Set project path prefix
proj_path <- "C:/Users/aruni/arunima/IISERTpt/Connectivity/"

# Read land cover rasters
lc_2017 <- rast(paste0(proj_path,"GIS/Land cover/WG Landscape change raster layers/2017d.img"))
lc_1995 <- rast(paste0(proj_path,"GIS/Land cover/WG Landscape change raster layers/1995d.img"))

# Drop unused categories
lc_2017 <- droplevels(lc_2017)
lc_1995 <- droplevels(lc_1995)

# Set background values to NA
lc_2017 <- subst(lc_2017,from = 0,to = NA,raw = TRUE)
lc_1995 <- subst(lc_1995,from = 0,to = NA,raw = TRUE)

# Name land cover categories
lc_df <- data.frame(value = 1:8,
                    class = c("Montane Grassland",
                              "Shola Forest",
                              "Timber Plantations",
                              "Tea Plantations",
                              "Ochlandra",
                              "Settlements",
                              "Agricultural Land",
                              "Water bodies"))

levels(lc_2017) <- lc_df
levels(lc_1995) <- lc_df

names(lc_2017) <- "landcov"
names(lc_1995) <- "landcov"

# Write processed raster to TIF file at original resolution
writeRaster(lc_2017,
            filename = paste0(proj_path,"GIS/Land cover/2017D/2017D_rast_30m.tif"),
            overwrite = TRUE)
writeRaster(lc_1995,
            filename = paste0(proj_path,"GIS/Land cover/1995D/1995D_rast_30m.tif"),
            overwrite = TRUE)

# Read reference 1ha resolution raster
rast_1ha <- rast(paste0(proj_path,"occupancy data/1500 1ha grids/1ha grids.tif"))

# Project land cover rasters to 1ha resolution and write to TIF files
lc_2017_1ha <- project(x = lc_2017,
                       y = rast_1ha,
                       method = "near",
                       filename = paste0(proj_path,"GIS/Land cover/2017D/2017D_rast_1ha.tif"),
                       overwrite = TRUE)
lc_1995_1ha <- project(x = lc_1995,
                       y = rast_1ha,
                       method = "near",
                       filename = paste0(proj_path,"GIS/Land cover/1995D/1995D_rast_1ha.tif"),
                       overwrite = TRUE)

# Polygonise rasters
lc_2017_vect <- as.polygons(lc_2017,round = FALSE)
lc_1995_vect <- as.polygons(lc_1995,round = FALSE)

# Save polygonised land cover shapefiles to directory
if (! dir.exists(paste0(proj_path,"GIS/Land cover/2017D/2017D vectorised"))) {
  dir.create(paste0(proj_path,"GIS/Land cover/2017D/2017D vectorised"),recursive = TRUE)
}

if (! dir.exists(paste0(proj_path,"GIS/Land cover/1995D/1995D vectorised"))) {
  dir.create(paste0(proj_path,"GIS/Land cover/1995D/1995D vectorised"),recursive = TRUE)
}

writeVector(lc_2017_vect,
            filename = paste0(proj_path,"GIS/Land cover/2017D/2017D vectorised/2017D.shp"),
            overwrite = TRUE)

writeVector(lc_1995_vect,
            filename = paste0(proj_path,"GIS/Land cover/1995D/1995D vectorised/1995D.shp"),
            overwrite = TRUE)

# # Clear environment and unused memory space
# rm(list = ls())
# gc()
# 
# # Set project path prefix
# proj_path <- "C:/Users/aruni/arunima/IISERTpt/Connectivity/"
# 
# # Read land cover rasters and polygonised land cover
# lc_2017_1ha <- rast(paste0(proj_path,"GIS/Land cover/2017D/2017D_rast_1ha.tif"))
# lc_1995_1ha <- rast(paste0(proj_path,"GIS/Land cover/1995D/1995D_rast_1ha.tif"))
# lc_2017_vect <- vect(paste0(proj_path,"GIS/Land cover/2017D/2017D vectorised/2017D.shp"))
# lc_1995_vect <- vect(paste0(proj_path,"GIS/Land cover/1995D/1995D vectorised/1995D.shp"))

# Create loop to derive percentage cover for each land cover class with all rasters
for (year in c("1995","2017")) {
  if (year == "1995") {
    lc_vect <- lc_1995_vect
    lc_grid <- lc_1995_1ha
  }
  
  if (year == "2017") {
    lc_vect <- lc_2017_vect
    lc_grid <- lc_2017_1ha
  }
  
  # Get fraction cover for each land cover class
  cover <- rasterize(lc_vect,
                     lc_grid,
                     field = "landcov",
                     fun = "mean",
                     cover = TRUE,
                     update = TRUE,
                     by = "landcov")
  
  # Rename layers
  names(cover) <- lapply(names(cover),function(x) {
    x <- gsub(" ","_",x)
    x <- paste0(x,"_cover")
    x
  })
  
  # Convert to percentage
  cover <- cover * 100
  
  # Set background pixels to 0
  cover[is.na(cover)] <- 0
  
  # Write rasters to TIF files
  output_paths <- c()
  
  if (! dir.exists(paste0(proj_path,"GIS/Derived rasters/Land cover/cover_",year,"_1ha"))) {
    dir.create(paste0(proj_path,"GIS/Derived rasters/Land cover/cover_",year,"_1ha"),recursive = TRUE)
  }
  
  for (i in names(cover)){
    output_paths <- c(output_paths,paste0(proj_path,"GIS/Derived rasters/Land cover/cover_",year,"_1ha/",i,".tif"))
  }
  
  writeRaster(cover,filename = output_paths,overwrite = TRUE)
}

# Extract Shola Forest and Timber plantation polygons
lc_2017_woodland <- lc_2017_vect[lc_2017_vect$landcov %in% c("Shola Forest","Timber Plantations")]
lc_1995_woodland <- lc_1995_vect[lc_1995_vect$landcov %in% c("Shola Forest","Timber Plantations")]

# Derive percentage woodland cover similarly
wland_2017_1ha <- rasterize(lc_2017_woodland,
                            lc_2017_1ha,
                            field = "landcov",
                            fun = "mean",
                            cover = TRUE,
                            update = TRUE)

wland_1995_1ha <- rasterize(lc_1995_woodland,
                            lc_1995_1ha,
                            field = "landcov",
                            fun = "mean",
                            cover = TRUE,
                            update = TRUE)

wland_2017_1ha <- wland_2017_1ha * 100
wland_1995_1ha <- wland_1995_1ha * 100

wland_2017_1ha[is.na(wland_2017_1ha)] <- 0
wland_1995_1ha[is.na(wland_1995_1ha)] <- 0

names(wland_2017_1ha) <- "Woodland_cover"
names(wland_1995_1ha) <- "Woodland_cover"

writeRaster(wland_2017_1ha,
            filename = paste0(proj_path,"GIS/Derived rasters/Land cover/cover_2017_1ha/Woodland_cover.tif"),
            overwrite = TRUE)
writeRaster(wland_1995_1ha,
            filename = paste0(proj_path,"GIS/Derived rasters/Land cover/cover_1995_1ha/Woodland_cover.tif"),
            overwrite = TRUE)
